Fix api select fn for safaridriver
Webdriver implementations seem to handle a click directly on the option
element without first click on the select element.

Except for `safaridriver`. It needs us to click on the select element
before clicking on the option element.

This seems like a reasonable general flow, so did not special case for
`safaridriver`.

Closes #403