Window size test now makes window smaller
The test was attempting to make the Window bigger, but for firefox on macOS on
GitHub Actions, this was failing. GitHub Actions macOS has a relatively
small screen size, and firefox won't resize larger if we are at the
limits.

I think we are still covering the spirit of the test if we resize
smaller instead of bigger.

Fixes #419