road to bb: remove dependency on apache commons.io
Etaoin was using apache commons.io to delete a tempoarary dir.
This dependency was brought in implicitly by clj-http.

Replaced with babashka/fs delete-tree.

Out of scope: Using more babashka/fs than we need to. This libary
has some great abstractions that I expect will help with maintainability,
we'll get to this later.

Contributes to #380