Switch from lein to clojure tools cli
Apologies to any lein lovers out there!

Reference to lein either converted to clojure cli or deleted.

- Figuring out how to call the official clojure distribution on Windows
is an exercise in frustration. Instead we use babashka's clojure function to
launch clojure. Works like a charm.
- Added new bb task to download clojure deps for CI
- Added an anemic pom.xml to preserve data previously held by project.clj.
This will be used by release flow once I get to #422.
- Unlike lein, the clojure test-runner does not allow selection of tests
based on metadata at the namespace level. To easily distinguish unit tests I
moved them under a `unit` directory.
- I noticed that ide tests were failing for me.
This is because we are now testing with our documented minimum Clojure
version of 1.9. The ide code was taking advantage of a 1.10 feature.
Making it 1.9 compatible was an easy fix.
- Switched to logback for logging for dev and testing to avoid any log4j
security concern stink.
- Noticed that we had `./resources/` `./env/dev/resources` and
`./env/test/resources`. But `resources/` contains only test resources, so I
moved it under `./env/test/resources`.  `./env/dev/resources` seems to
be a way to turn on debug logging, so left it as that and created a
deps.edn `:debug` alias for it.
- Our Makefile is getting smaller. Docker tasks now certainly broken.
I'll make a separate issue to deal with this (and move to bb tasks).

Helps me with #380
Closes #432