Declare the etaoin public API
Clearly declaring a public API allows us to freely change what is
internal and invest our effort in being careful to not break what is
public.

We've asked the Etaoin community if we've got this right, but no peep
yet. I'm sure we'll hear plenty of peeps if we've got it wrong when
folks actually start using the next release. If so, we can adjust as
necessary.

Moved what namespaces are deemed internal to impl namespaces and marked
with `:no-doc` metadata.

See #430 for details.

Closes #430