Include driver process liveness in some exceptions
When an http request, for example, times out one might wonder if the
WebDriver process is still running.

We now use slingshot throw+ as we do for http status errors. This throw+
still includes the :driver :process but checks if it is alive and
realizes its exit value in the thrown map if dead.

Also: if an exception occurs when booting a driver, any launched WebDriver
process is now cleaned up.

Also: reworked some proc tests to always attempt cleanup even when
exceptions occur.

Closes #469