docs: user guide chord examples now os-agnostic
Keyboard chord examples formerly macOS specific.
Add Ubuntu and Windows for test-doc to CI matrix.

Weirdness: when testing locally on Linux, for sessions with no real
interaction, was sporadically getting socket timeout when trying to quit
session. For now, introduce a couple of waits in doc examples to appease
the mystery.

I expect I'll need to tweak a thing or two to get this working on CI.
Do I need to setup a virtual display for Ubuntu?