ci: fix for safari
It seems that GitHub Actions no longer supports `file:` urls for when
using safaridriver. I spent some time trying to coax it to do so but
safaridriver is not well documented, so instead of trying to come up
with brittle incantations, now hitting a local http test server for api
tests instead.

Cookie tests were affected by this changed and adjusted/simplified:
- browser specific tests consolidated (and edge now included)
- turfed phantom test, we no longer support it

Fixes #561