ci: adapt to new macos arm-based runners
The new macos runner is arm-based and missing some software is missing
that all other runners have:
- ImageMagick
- Edge and msedgedriver
- Firefox and geckodriver
Take a shot at optionally installing the above when we need them.

There is also no JDK8 for arm-based macos.
We weren't testing that particular scenario on the intel-based macos runners
but I updated the code to reflect this fact.