Minimal safari driver logging (#601)
* Minimal safari driver logging

Safaridriver logging is relatively odd. It only has debug level logging,
only logs to a file, and discovering that file is somewhat tricky.

This change adds minimal support to discover and dump safaridriver logs.

I'm not sure how many folks bother with safaridriver testing, so this
change might be mostly a favour to myself when diagnosing safaridriver
issues that arise from time to time.

Closes #563
Incidentally fixes #517

* Oh, Windows does not have SIGTERM (or signals)

My initial approach to logging termination of our fake-driver was to
have it listen for SIGTERM. This works on macOS/Linux but Windows has
no facility to listen for termination signals. So instead, we hook into
`:post-stop-fns` for something that works for all OSes.