ci: publish: Move to release tag triggered publishing (#611)
* Move to release tag triggered publishing

I still prefer my old GitHub UI triggered release. It feels much less
complicated to me. But this project is part of clj-commons, and here
I defer to the larger team's preference for release triggered publishing.

I brought over this work from other projects, mostly rewrite-clj and
pomegranate. These incude:

- using neil to track/bump/store version in deps.edn
- automatically including date in changelog release headers

One refinement for Etaoin is avoiding the duplicate test run on
Publish. The release tag triggered strategy includes a commit and tag
from the person doing the publish. This commit triggers the normal
Test workflow but the Publish workflow also triggers the Test workflow
to ensure all is good before releasing. To avoid the duplicate test run,
I've added a new `check-for-skip` job in the Test workflow to skip tests
when appropriate. I'd love to skip the independent Test workflow on Publish
entirely, but don't think the GitHub Actions YAML magic to do so exists.

Closes #529

* merge new check-for-skip job into setup job

Since the check-for-skip work needs pretty much the same setup as the
setup job, merge the two and avoid the extra job.

* yaml syntax

* fix yaml

* [publish workflow] should cause tests to be skipped

A wee sanity test

* [publish workflow] debug commit msg

* remove println debug

On a PR the commit message is the merge commit.
That's not important to me, I care about behaviour on master, the branch
from which we publish.