Issue 604 addendum (#619)
* Change shadow root query function names (#604)

Update the query function names for shadow roots to be closer to the
API names used in the Web Driver spec.

* Rework get-element-shadow-root(-el) to use Web Driver APIs

Shifts from using get-element-property-el to using a new
get-element-shadow-root* based on the underlying Web Driver Get
Element Shadow Root API endpoint.

* Use unwrap-webdriver-object for find-element(s)-from-shadow-root*

Previous find-element(s)-from-shadow-root* functions used the first /
second idiom to extract the value of the first pair in a map. This is
brittle and may not work in the future for some reason. This is
replaced with unwrap-webdriver-object looking for a specific item type
tag in the map, as specified by the Web Driver spec.

* Update get-element-shadow-root* to use get-element-property-el

Originally, tried using Web Driver Get Element Shadow Root API call,
but the error handling for this was too inconsistent across multiple
browers. Using get-element-property-el works much more consistently
and avoids having to deal with exception processing as much.

* Update shadow-root functions to work with query's vector syntax

* woops, missed committing something here!

* address lint error from eastwood

* Don't need `:ELEMENT` from chrome/edge anymore

Now that we are in "w3c mode" for chrome and edge we do not need to
special case them anymore.

* Use new `web-element-identifier` constant

Use this new constant to replace magic number in existing code.

* Add date to comment on WebDriver behaviours

To make it easy to see when this was last checked.

---------

Co-authored-by: lread <lee@dlread.com>