`:no-doc` to `:private` (#651)
* Use locator-xpath and locator-css in etaoin/query

locator-xpath and locator-css were duplicated in both etaoin/api and
etaoin/query, but the versions in etaoin/api were marked as
`:no-doc`. This change prefers the versions in etaoin/query and uses
those throughout the code base, removing the definitions from
etaoin/api.

* Make dispatch-driver private

Also modify a macro and a test that used dispatch-driver to instead
use driver-type, which will be a public API.

* Update CHANGELOG with issue 633 information

* Revert CHANGELOG update

The changes a private and non-breaking, so no need to announce them.