Make vector arg to fill-multi/fill-human-multi fill in order (#652)
* Make vector arg to fill-multi/fill-human-multi fill in order

Previously, if the user supplied a vector argument to fill-multi or
fill-human-multi, Etaoin would convert the vector to a map, then
convert the map to a sequence of MapEntry pairs, and use those to fill
multiple fields. Unfortunately, this threw away all the ordering
information in the vector. This change processes the vector as a
sequence of partitioned pairs and thereby keeps the order intact.

* Move CHANGELOG entries to correct section

* Move CHANGELOG entries into correct sections

* Remove "Docs" heading from CHANGELOG

* Ordered CHANGELOG entries numerically